Player data loading, follow camera and player collision detection
Implemented the ability for player save data to be loaded from file and populate the player register. Implemented a version of the following camera, may change to point to a specific entity as opposed to always being the player. Implemented the player-to-world collision detection and response (May re-write this due to it being long-winded).