Added texture picking by file and Entity Type/Colour via Dropdown box to prevent errors
There is now two config files in the content folder which store the entity types available and the colours viable for rendering/conversion within the engine. The Add Item/Entity now also use a file picker dialog to choose there texture tags, instead of the old way of manually typing which may cause errors if the texture tag was mispelt etc.