Reimplemented wall snapping
Fixed wall snapping to be far more efficient, should now work for any speed up to 960, which is where the player will be able to move more than a tile per frame so collisions will be missed. Slight issue found with wall sliding where if trying to slide along tiles with bounding boxes which do not fill the whole 16x16 area.