Reimplemented the camera, seems to be working
Reimplemented the camera to use lerp and better use of transformation matrices. Needs further testing for the between tile artefacts which were occuring. Needs capping to the level/room bounds and the IsInView implemented. Do the capping through manipulating the viewport offset depending on how many tiles are left before the bounds of the level/room. Do the IsInView by working out what 'room' of the world the player is in and BSPing the world down for render sections