Cleaned Repo and Re-implemented RenderTarget2D
Re-implemented the use of RenderTarget2D because the way it works is beneifical to future expansions of how the renderer works. Black line issue appears to have been caused by the camera lerp reaching strange values which messed with textures, implement a Movetowards() for the camera and see what happens, old lerp didn't work and was insanely broken. If the Movetowards() doesn't work, try adding an extrude to all textures.