Camera work and animation work
Camera: Implemented the camera being clamped to the bounds of the world (ready for expansion to clamped to rooms, once the rooms have been implemented). Animations: implemented the animations directions being detached from the directiokn the agent is moving, it now uses the velocity to decide, this prevents the weird change in animations which where occuring, only supports four directions right now,will consider eight. The objects now use a delegate for their animationhandler and animationfinders so they can be assigned in derived classes without needing overriding and allows tiles to now use their own way of animating.